requirements.txt
streamlit
opencv-python-headless
mediapipe
pillow
numpy
torch
torchvision
